Add description for Pork giniling
Pork Giniling is a pork dish that makes use of ground pork as the primary ingredient. Ground pork is stewed in tomato sauce and water to bring out the taste while vegetables such as carrots and potatoes (some also like this with raisins and green peas) are added for additional flavor and nutrition.