Add Instructions on how to cook Pork Giniling
Heat a cooking pot and pour-in the cooking oil.
When the oil is hot enough, put-in the garlic and sauté until the color turns light brown.
Add the onions and sauté until the texture becomes soft. Put-in the ground pork and cook for 5 minutes.
Put-in the ground pork and cook for 5 minutes.
Add the beef or pork cube, tomato sauce, and water and let boil. Simmer for 20 minutes.
Put the carrots and potatoes in then stir until every ingredient is properly distributed. Simmer for 10 to 12 minutes.
Add salt, ground black pepper, and sugar then stir.
Put in the boiled eggs and turn off the heat.
Transfer to a serving bowl and serve.
Share and enjoy!